Move all devnet client tests to network tests
