+++ docs/development.rst
@@ -23,7 +23,7 @@ To install `starknet-devnet-rs <https://github.com/0xSpaceShard/starknet-devnet-
 Environment variables
 ^^^^^^^^^^^^^^^^^^^^^
 
-To run tests on testnet network (``starknet_py/tests/e2e/tests_on_networks/``) additionally you must set:
+In order to be able to run tests on testnet network (``starknet_py/tests/e2e/tests_on_networks/``), you must set some environmental variables:
 
     - ``SEPOLIA_RPC_URL``
     - ``SEPOLIA_ACCOUNT_PRIVATE_KEY``
