+++ starknet_py/tests/e2e/tests_on_networks/client_devnet/fixtures/contracts.py
